Deprecate Classic Mode (Part 1)
  Related to #2548

  - Replace all occurrences of AppRegistry by AppRegistryCommon or AppRegistyrService. This moves all app registration (including tasks) from UriAppRegistry to DefaultAppRegistryService
  - The AbstractShellIntegrationTest and therefore all CommandsXXX tests spin an embedded SCDF server - Had to mock the skipperClient and add default mode About response
  - In DataFlowControllerAutoConfiguration#skipperClient bean factory remove the Jackson2HalModule() (see https://pivotal.slack.com/archives/G49UPQAP8/p1541431130391100 and https://pivotal.slack.com/archives/C055XJPLR/p1540810022013200 )
  - Add optional StreamDefinitionRepository and StreamService dependencies for SkipperAppRegistryController. Later are used for app unregister validation but are not present for stream feature disabled!  It is awkward that Apps should depend on Streams. This dependencies should be other way around.  This also manifestations the split app state, split between SCDF and Skipper
  - Remove all Classic Mode bean definitions from DataFlowControllerAutoConfiguration and TestDependecies
  - Remove Skipper feature and all its occurrences and checks. Now all previous skipper related configurations are subject of the Stream Feature policies.
  - The StreamControllerTests#testDeployWithWildcardProperties demonstrates how to use the skipperClient mock and UploadRequest and InstallRequest to verify the Stream Creation and Deployment behavior without relying on local app deployer
  - Removed classes:
     AppDeployerStreamDeployer, AppDeployerStreamDeployerTests
     AppDeployerStreamService
     AppDeployerStreamServiceTests
     AppRegistry, AppRegistryTests
     AppRegistryController, AppRegistryControllerTests
     ClassicAppRegistryCommands, ClassicAppRegistryCommandsTests, ClassicStreamCommands
     ConditionalOnSkipperDisabled, ConditionalOnSkipperEnabled
     DataFlowMode, IncompatibleDataFlowMode
     org.springframework.cloud.dataflow.shell.command.classic Package
     RdbmsUriRegistry, RdbmsUriRegistryTests
     RedisUriRegistry

- Resolve Jackson2HalModule configuration
- Remove redundant,  classic-mode classes.
- Improve test coverage.
- mock skipper client for tests

+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/controller/SkipperAppRegistryController.java
@@ -25,6 +25,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Properties;
 import java.util.concurrent.ForkJoinPool;
 
@@ -103,13 +104,13 @@ public class SkipperAppRegistryController {
 
 	private ResourceLoader resourceLoader = new DefaultResourceLoader();
 
-	public SkipperAppRegistryController(StreamDefinitionRepository streamDefinitionRepository,
-			StreamService streamService,
+	public SkipperAppRegistryController(Optional<StreamDefinitionRepository> streamDefinitionRepository,
+			Optional<StreamService> streamService,
 			AppRegistryService appRegistryService,
 			ApplicationConfigurationMetadataResolver metadataResolver,
 			ForkJoinPool forkJoinPool, MavenProperties mavenProperties) {
-		this.streamDefinitionRepository = streamDefinitionRepository;
-		this.streamService = streamService;
+		this.streamDefinitionRepository = streamDefinitionRepository.isPresent() ? streamDefinitionRepository.get() : null;
+		this.streamService = streamService.isPresent() ? streamService.get() : null;
 		this.appRegistryService = appRegistryService;
 		this.metadataResolver = metadataResolver;
 		this.forkJoinPool = forkJoinPool;
@@ -166,9 +167,7 @@ public class SkipperAppRegistryController {
 		if (!this.appRegistryService.appExist(name, type)) {
 			throw new NoSuchAppRegistrationException(name, type);
 		}
-		if (this.appRegistryService.getDefaultApp(name, type) == null) {
-			throw new RuntimeException(String.format("No default version exists for the app [%s:%s]", name, type));
-		}
+
 		String defaultVersion = this.appRegistryService.getDefaultApp(name, type).getVersion();
 		return getInfo(type, name, defaultVersion, exhaustive);
 	}
@@ -265,6 +264,10 @@ public class SkipperAppRegistryController {
 			}
 		}
 
+		if (!this.appRegistryService.appExist(name, type, version)) {
+			throw new NoSuchAppRegistrationException(name, type, version);
+		}
+
 		appRegistryService.delete(name, type, version);
 	}
 
@@ -278,12 +281,15 @@ public class SkipperAppRegistryController {
 	 * return {@code null}.
 	 */
 	private String findStreamContainingAppOf(ApplicationType appType, String appName, String appVersion) {
+		if (this.streamDefinitionRepository == null || this.streamService == null) {
+			return null;
+		}
 		Iterable<StreamDefinition> streamDefinitions = streamDefinitionRepository.findAll();
 		for (StreamDefinition streamDefinition : streamDefinitions) {
 			StreamDeployment streamDeployment = this.streamService.info(streamDefinition.getName());
 			for (StreamAppDefinition streamAppDefinition : streamDefinition.getAppDefinitions()) {
 				final String streamAppName = streamAppDefinition.getRegisteredAppName();
-				final ApplicationType streamAppType =streamAppDefinition.getApplicationType();
+				final ApplicationType streamAppType = streamAppDefinition.getApplicationType();
 				if (appType != streamAppType) {
 					continue;
 				}
